e2e/src/app.e2e-spec.ts
import { browser, by, element } from 'protractor';

describe('workspace-project App', () => {
  beforeAll(() => {
    // Configuración inicial si es necesario
  });

  it('should display welcome message', () => {
    browser.get('/');
    element(by.css('ion-title')).getText().then((text) => {
      expect(text).toEqual('Welcome to skeleton-app!');
    });
  });

  it('should access the camera', () => {
    // Simulación de acceso a la cámara
    const cameraButton = element(by.css('.camera-button'));
    cameraButton.click();
    // Aquí deberías agregar la lógica para verificar que la cámara se ha abierto correctamente
    // Esto puede variar dependiendo de cómo esté implementada la funcionalidad de la cámara
    element(by.css('.camera-view')).isPresent().then((isPresent) => {
      expect(isPresent).toBe(true);
    });
  });

  it('should access the GPS', () => {
    // Simulación de acceso al GPS
    const gpsButton = element(by.css('.gps-button'));
    gpsButton.click();
    // Aquí deberías agregar la lógica para verificar que el GPS se ha activado correctamente
    // Esto puede variar dependiendo de cómo esté implementada la funcionalidad del GPS
    element(by.css('.gps-coordinates')).getText().then((coordinates) => {
      expect(coordinates).toMatch(/Latitude: \d+\.\d+, Longitude: \d+\.\d+/);
    });
  });

  afterAll(() => {
    // Limpieza si es necesario
  });
});

describe('Pruebas End-To-End', () => {
  it('debería tener un título', () => {
    browser.get('/');
    browser.getTitle().then((title) => {
      expect(title).toEqual('Mi Aplicación');
    });
  });

  // ...otras pruebas...
});